+++ source/js/search-parameters/search-parameters.js
@@ -1,22 +1,39 @@
-import { getAutocompleterById, toggleCssClass } from '../common/utils';
+import {
+  getAutocompleterById,
+  getFocusableChildren,
+  toggleCssClass
+} from '../common/utils';
 import {
   getSearchParamGroupFactoryByResourceType,
   setFhirServerForSearchParameters
 } from './common-descriptions';
+import { BaseComponent } from '../common/base-component';
 
-export class SearchParameters {
+/**
+ * Component class for managing criteria
+ */
+export class SearchParameters extends BaseComponent {
   /**
-   * Inserts the component after table row selected by anchorSelector
-   * @param {string} anchorSelector
-   * @param {string} serviceBaseUrl - FHIR REST API Service Base URL (https://www.hl7.org/fhir/http.html#root)
-   * @param {Object[]} searchParamGroups Array of objects describing the search parameters
+   * Constructor of component
+   * @param {Object<Function>} callbacks - callback functions that the component uses for input/output
+   * @param {string} serviceBaseUrl - FHIR REST API Service Base URL (https://www.hl7.org/fhir/http.html#root).
+   *        On instantiating the class, the metadata for the specified server will be loaded and used to
+   *        determine the FHIR version. The FHIR version is needed to get the specification of the criteria.
+   *        Use "ready" property (a Promise) to check that the component is ready for use.
+   * @param {Object[]} searchParamGroups - array of objects describing the search parameters
    *                   (see patient-search-parameters.js for an example object)
    *                   or strings with resource types.
    */
-  constructor(anchorSelector, serviceBaseUrl, searchParamGroups) {
-    this.initialize = setFhirServerForSearchParameters(serviceBaseUrl).then(
+  constructor({ callbacks, serviceBaseUrl, searchParamGroups }) {
+    super({
+      callbacks
+    });
+    /**
+     * This promise will be resolved when component is ready to use
+     * @type {Promise<void>}
+     */
+    this.ready = setFhirServerForSearchParameters(serviceBaseUrl).then(
       () => {
-        this.internalId = 'searchParam';
         this.availableParams = {};
         this.searchParams = searchParamGroups.reduce((_searchParams, item) => {
           const searchParamGroupFactory =
@@ -33,13 +50,13 @@ export class SearchParameters {
             : Object.keys(searchParamGroup.description).sort();
           return _searchParams;
         }, {});
-        this.buttonId = this.internalId + '_add_button';
-        this.item_prefix = this.internalId + '_param_';
+        this.buttonId = this._id + '_add_button';
+        this.item_prefix = this._id + '_param_';
         this.item_generator = 0;
         this.selectedParams = {};
         this.selectedResources = {};
 
-        this.addButton(anchorSelector);
+        this.initialize();
 
         Def.Autocompleter.Event.observeListSelections(null, (eventData) => {
           const inputId = eventData.field_id;
@@ -108,17 +125,54 @@ export class SearchParameters {
   }
 
   /**
-   * Returns root HTMLElement of the SearchParameters component
-   * @return {HTMLElement}
+   * Returns HTML for component
+   * @return {string}
    */
-  getHtmlElement() {
-    return document.getElementById(this.internalId);
+  getHtml() {
+    return `\
+<div id="${this._id}" class="search-parameter-list${
+      this.getAvailableResourceTypes().length === 1
+        ? ' search-parameter-list_one-resource'
+        : ''
+    } search-parameter-list_empty">
+    <div class="search-parameter-list__combiner"><label>AND</label> - criteria are combined with logical AND</div>
+    <div class="section__body"></div>
+</div>
+<div>
+  <button id="${
+    this.buttonId
+  }" type="button" class="add-search-param-button">Add a search criterion</button>
+</div>`;
+  }
+
+  /**
+   * Initializes controls created in getHtml
+   */
+  attachControls() {
+    this.attachEvent(document.getElementById(this.buttonId), 'click', () =>
+      this.addParam(false)
+    );
+    this.attachEvent(
+      document.getElementById(this.buttonId),
+      'keypress',
+      (event) => {
+        if (
+          event.key === ' ' ||
+          event.key === 'Spacebar' ||
+          event.key === 'Enter'
+        ) {
+          this.addParam(true);
+          event.preventDefault();
+        }
+      }
+    );
   }
 
   /**
    * Removes all controls and related data
    */
-  dispose() {
+  detachControls() {
+    super.detachControls();
     const addBtn = document.getElementById(this.buttonId);
     if (addBtn) {
       Object.keys(this.selectedParams).forEach((searchItemId) =>
@@ -331,30 +385,11 @@ export class SearchParameters {
   }
 
   /**
-   * Adds a button to add search parameters
-   * @param {string} anchorSelector table row selector after which a row with a button will be added
-   */
-  addButton(anchorSelector) {
-    const anchorElement = document.querySelector(anchorSelector);
-
-    anchorElement.insertAdjacentHTML(
-      'afterend',
-      `\
-<div id="${this.internalId}" class="search-parameter-list search-parameter-list_empty">
-    <div class="search-parameter-list__combiner"><label>AND</label> - criteria are combined with logical AND</div>
-    <div class="section__body"></div>
-</div>
-<div>
-  <button id="${this.buttonId}" type="button" class="add-search-param-button">Add a search criterion</button>
-</div>`
-    );
-    document.getElementById(this.buttonId).onclick = () => this.addParam();
-  }
-
-  /**
    * Adds a new row with search parameter to the table of search parameters
+   * @param {boolean} accessibilityMode - focus moves to the first field of
+   *        the new criterion
    */
-  addParam() {
+  addParam(accessibilityMode) {
     // searchItemId is the unique virtual identifier of the controls group for each search parameter.
     // Used as part of HTML element identifiers and for storing data associated with a search parameter.
     const searchItemId = this.item_prefix + ++this.item_generator;
@@ -365,7 +400,7 @@ export class SearchParameters {
     const searchItemContentId = this.getParamContentId(searchItemId);
     const removeButtonId = this.getRemoveButtonId(searchItemId);
     const prevRowElement = document
-      .getElementById(this.internalId)
+      .getElementById(this._id)
       .querySelector('.section__body >:last-child');
     const prevResourceTypeSelector = prevRowElement
       ? document.getElementById(
@@ -386,18 +421,18 @@ export class SearchParameters {
     this.selectedResources[searchItemId] = paramResourceType;
 
     document
-      .getElementById(this.internalId)
+      .getElementById(this._id)
       .querySelector('.section__body')
       .insertAdjacentHTML(
         'beforeend',
         `\
 <div id="${rowId}" class="${this.getSearchParameterClass(searchItemId)}">
-  <input type="text" id="${paramResourceTypeSelectorId}" value="${paramResourceType}">
+  <input type="text" id="${paramResourceTypeSelectorId}" value="${paramResourceType}" aria-label="Resource type">
   <div class="search-parameter__name">
-    <input type="text" id="${searchItemId}" value="${paramName}">
+    <input type="text" id="${searchItemId}" value="${paramName}" aria-label="Search parameter name">
   </div>
   <div id="${searchItemContentId}" class="search-parameter__content"></div>
-  <button id="${removeButtonId}" type="button">remove</button>
+  <button id="${removeButtonId}" type="button" aria-label="Remove the search criterion before this button">remove</button>
 </div>`
       );
     new Def.Autocompleter.Prefetch(paramResourceTypeSelectorId, [], {
@@ -409,6 +444,17 @@ export class SearchParameters {
     this.addRemoveBtnListener(searchItemId);
     this.onParamsCountChanged();
     this.createControlsForSearchParam(searchItemId);
+
+    if (accessibilityMode) {
+      // Focus moves to the first field of the new criterion
+      const focusableElements = getFocusableChildren(
+        document.getElementById(rowId)
+      );
+      if (focusableElements.length > 0) {
+        focusableElements[0].focus();
+      }
+    }
+
     if (!this.getAvailableResourceTypes().length) {
       document.getElementById(this.buttonId).disabled = true;
     }
@@ -459,7 +505,7 @@ export class SearchParameters {
   onParamsCountChanged() {
     this.updateAllSearchParamSelectors();
     const paramsCount = Object.keys(this.selectedParams).length;
-    const paramListElement = document.getElementById(this.internalId);
+    const paramListElement = document.getElementById(this._id);
     toggleCssClass(
       paramListElement,
       'search-parameter-list_empty',